+++ lib/model/Groups.dart
@@ -825,9 +825,10 @@ class GroupPost {
   final DateTime? dateCreatedUtc;
   final DateTime? dateUpdatedUtc;
   final bool? private;
+  final String? imageUrl;
   final List<GroupPost>? replies;
 
-  GroupPost({this.id, this.parentId, this.member, this.subject, this.body, this.dateCreatedUtc, this.dateUpdatedUtc, this.private, this.replies});
+  GroupPost({this.id, this.parentId, this.member, this.subject, this.body, this.dateCreatedUtc, this.dateUpdatedUtc, this.private, this.imageUrl, this.replies});
 
   static GroupPost? fromJson(Map<String, dynamic>? json) {
     return (json != null) ? GroupPost(
@@ -839,8 +840,8 @@ class GroupPost {
         dateCreatedUtc: groupUtcDateTimeFromString(json['date_created']),
         dateUpdatedUtc: groupUtcDateTimeFromString(json['date_updated']),
         private: json['private'],
-        replies: GroupPost.fromJsonList(json['replies'])
-      ) : null;
+        imageUrl: AppJson.stringValue(json["image_url"]),
+        replies: GroupPost.fromJsonList(json['replies'])) : null;
   }
 
   Map<String, dynamic> toJson({bool create = false, bool update = false}) {
@@ -854,6 +855,9 @@ class GroupPost {
     }
     if (subject != null) {
       json['subject'] = subject;
+    }
+    if(imageUrl!=null){
+      json['image_url'] = imageUrl;
     }
     return json;
   }